Added 64-bit support.

+++ BTRatingView/BTRatingView.m
@@ -161,7 +161,7 @@
     if (!self.editable) return;
     
     NSInteger newRating = 0;
-    for (int i = self.imageViews.count - 1; i >= 0; i--) {
+    for (NSInteger i = self.imageViews.count - 1; i >= 0; i--) {
         UIImageView *imageView = [self.imageViews objectAtIndex:i];
         if (touchLocation.x > imageView.frame.origin.x) {
             newRating = i+1;
